+++ packages/opentelemetry-exporter-collector/src/platform/browser/CollectorTraceExporter.ts
