+++ packages/opentelemetry-exporter-collector/src/platform/browser/index.ts
@@ -14,5 +14,5 @@
  * limitations under the License.
  */
 
-export * from './CollectorExporter';
+export * from './CollectorTraceExporter';
 export * from './CollectorMetricExporter';
